+++ src/ButtonIcon/stories.js
@@ -1,7 +1,7 @@
 import React from 'react'
 import { storiesOf } from '@storybook/react'
 import { action } from '@storybook/addon-actions'
-import { radios } from '@storybook/addon-knobs'
+import { radios, boolean } from '@storybook/addon-knobs'
 
 import Tag from '../Tag'
 import Icon from '../Icon'
@@ -12,6 +12,8 @@ storiesOf('ButtonIcon')
     <figure>
       <ButtonIcon
         onClick={action('onClick event')}
+        disabled={boolean('Disabled', false)}
+        busy={boolean('Busy', false)}
         variant={radios('Variant', ['primary', 'secondary', 'destructive', 'chevron'], 'primary')}>
         <Icon name="user" />
       </ButtonIcon>